requirements.txt
fastapi
uvicorn
yt-dlp
python-dotenv
pydantic